+++ ibllib/atlas/atlas.py
@@ -220,27 +220,6 @@ class BrainAtlas:
             self.srf_xyz = self.bc.i2xyz(np.c_[idx_srf[self.xyz2dims[0]], idx_srf[self.xyz2dims[1]],
                                                idx_srf[self.xyz2dims[2]]].astype(float))
 
-    def compute_boundaries(self):
-        """
-        Compute the boundaries between regions
-        """
-        if self.boundary is None:  # only compute if it hasn't already been computed
-            self.boundary = np.diff(self.label, axis=0, append=0)
-            self.boundary = self.boundary + np.diff(self.label, axis=1, append=0)
-            self.boundary = self.boundary + np.diff(self.label, axis=2, append=0)
-            self.boundary[self.boundary != 0] = 1
-
-            # Compute boundary on top view
-            self.compute_surface()
-            ix, iy = np.meshgrid(np.arange(self.bc.nx), np.arange(self.bc.ny))
-            iz = self.bc.z2i(self.top)
-            inds = self._lookup_inds(np.stack((ix, iy, iz), axis=-1))
-            vals = np.random.random(self.regions.acronym.shape[0])
-            _top_boundary = vals[self.label.flat[inds]]
-            self.top_boundary = np.diff(_top_boundary, axis=0, append=0)
-            self.top_boundary = self.top_boundary + np.diff(_top_boundary, axis=1, append=0)
-            self.top_boundary[self.top_boundary != 0] = 1
-
     def _lookup_inds(self, ixyz):
         """
         Performs a 3D lookup from volume indices ixyz to the image volume
@@ -467,12 +446,24 @@ class BrainAtlas:
             self.compute_surface()
             return _take(self.surface, index, axis=self.xyz2dims[axis])
         elif volume == 'boundary':
-            self.compute_boundaries()
-            return _take(self.boundary, index, axis=self.xyz2dims[axis])
+            iregion = _take_remap(self.label, index, self.xyz2dims[axis], mapping)
+            return self.compute_boundaries(iregion)
+
         elif volume == 'volume':
             if bc is not None:
                 index = bc.xyz2i(np.array([coordinate] * 3))[axis]
             return _take(region_values, index, axis=self.xyz2dims[axis])
+
+    def compute_boundaries(self, values):
+        """
+        Compute the boundaries between regions on slice
+        :param values:
+        :return:
+        """
+        boundary = np.diff(values, axis=0, append=0)
+        boundary = boundary + np.diff(values, axis=1, append=0)
+        boundary[boundary != 0] = 1
+        return boundary
 
     def plot_cslice(self, ap_coordinate, volume='image', mapping='Allen', region_values=None, **kwargs):
         """
@@ -496,7 +487,7 @@ class BrainAtlas:
         """
 
         cslice = self.slice(ap_coordinate, axis=1, volume=volume, mapping=mapping, region_values=region_values)
-        return self._plot_slice(cslice.T, extent=self.extent(axis=1), **kwargs)
+        return self._plot_slice(np.moveaxis(cslice, 0, 1), extent=self.extent(axis=1), **kwargs)
 
     def plot_hslice(self, dv_coordinate, volume='image', mapping='Allen', region_values=None, **kwargs):
         """
@@ -582,8 +573,7 @@ class BrainAtlas:
                 for y in range(im.shape[1]):
                     im[x, y] = region_values[x, y, iz[x, y]]
         elif volume == 'boundary':
-            self.compute_boundaries()
-            im = self.top_boundary
+            im = self.compute_boundaries(regions)
 
         return self._plot_slice(im, self.extent(axis=2), ax=ax, **kwargs)
 
@@ -982,4 +972,90 @@ def _download_atlas_allen(file_image, FLAT_IRON_ATLAS_REL_PATH, par):
         raise ValueError('Unrecognized file image')
 
     cache_dir = Path(par.CACHE_DIR).joinpath(FLAT_IRON_ATLAS_REL_PATH)
-    return http_download_file(url, cache_dir=cache_dir)
+    return http_download_file(url, target_dir=cache_dir)
+
+
+class FlatMap(AllenAtlas):
+
+    def __init__(self, flatmap='dorsal_cortex', res_um=25):
+        """
+        Avaiable flatmaps are currently 'dorsal_cortex', 'circles' and 'pyramid'
+        :param flatmap:
+        :param res_um:
+        """
+        super().__init__(res_um=res_um)
+        self.name = flatmap
+        if flatmap == 'dorsal_cortex':
+            self._get_flatmap_from_file()
+        elif flatmap == 'circles':
+            from ibllib.atlas.flatmaps import circles
+            if res_um != 25:
+                raise NotImplementedError('Pyramid circles not implemented for resolution other than 25um')
+            self.flatmap, self.ml_scale, self.ap_scale = circles(N=5, atlas=self, display='flat')
+        elif flatmap == 'pyramid':
+            from ibllib.atlas.flatmaps import circles
+            if res_um != 25:
+                raise NotImplementedError('Pyramid circles not implemented for resolution other than 25um')
+            self.flatmap, self.ml_scale, self.ap_scale = circles(N=5, atlas=self, display='pyramid')
+
+    def _get_flatmap_from_file(self):
+        # TODO probably best to store depth as the c order??
+        # gets the file in the ONE cache for the flatmap name in the property, downloads it if needed
+        par = one.params.get(silent=True)
+        FLAT_IRON_ATLAS_REL_PATH = PurePosixPath('histology', 'ATLAS', 'Needles', 'Allen', 'flatmaps')
+        path_atlas = Path(par.CACHE_DIR).joinpath(FLAT_IRON_ATLAS_REL_PATH)
+        file_flatmap = path_atlas.joinpath(f'{self.name}_{self.res_um}.nrrd')
+        if not file_flatmap.exists():
+            self._download_flatmap(file_flatmap, FLAT_IRON_ATLAS_REL_PATH, par)
+        self.flatmap, _ = nrrd.read(file_flatmap)
+
+    @staticmethod
+    def _download_flatmap(file_image, FLAT_IRON_ATLAS_REL_PATH, par):
+        """
+        Download flatmap volumes
+        :param file_image:
+        :param FLAT_IRON_ATLAS_REL_PATH:
+        :param par:
+        :return:
+        """
+        file_image.parent.mkdir(exist_ok=True, parents=True)
+        base_url = (par.HTTP_DATA_SERVER + '/histology/ATLAS/Needles/Allen/flatmaps')
+        flatmap_url = base_url + '/' + file_image.name
+
+        cache_dir = Path(par.CACHE_DIR).joinpath(FLAT_IRON_ATLAS_REL_PATH)
+        return http_download_file(flatmap_url, target_dir=cache_dir, username=par.HTTP_DATA_SERVER_LOGIN,
+                                  password=par.HTTP_DATA_SERVER_PWD)
+
+    def plot_flatmap(self, depth=0, volume='annotation', mapping='Allen', region_values=None, ax=None, **kwargs):
+        """
+        Displays the 2D image corresponding to the flatmap. If there are several depths, by default it
+        will display the first one
+        :param depth: index of the depth to display in the flatmap volume (the last dimension)
+        :param volume:
+        :param mapping:
+        :param region_values:
+        :param ax:
+        :param kwargs:
+        :return:
+        """
+        if self.flatmap.ndim == 3:
+            inds = np.int32(self.flatmap[:, :, depth])
+        else:
+            inds = np.int32(self.flatmap[:, :])
+        regions = self._get_mapping(mapping=mapping)[self.label.flat[inds]]
+        if volume == 'annotation':
+            im = self._label2rgb(regions)
+        elif volume == 'value':
+            im = region_values[regions]
+        elif volume == 'boundary':
+            im = self.compute_boundaries(regions)
+        elif volume == 'image':
+            im = self.image.flat[inds]
+        if not ax:
+            ax = plt.gca()
+
+        return self._plot_slice(im, self.extent_flmap(), ax=ax, **kwargs)
+
+    def extent_flmap(self):
+        extent = np.r_[0, self.flatmap.shape[1], 0, self.flatmap.shape[0]]
+        return extent