+++ release_notes.md
@@ -1,7 +1,18 @@
-## To be added to next release
+## Release Note 2.10
+
+## Release Note 2.10.0 2022-02-11
 - Fix in EphysDLC task to not return multiple copies of outputs
 - Loading examples for different IBL data types
 - Fix for probe syncing when Nidq and probe pulses don't match
+- Account for new ONE tables in ond datahandler
+- Add bad channels plots into RawEphysQc task
+- SpikeGlx:
+  - supports reading of flat binary files without meta data
+- atlas.Regions:
+  - add support for region ordering in BrainRegions object
+  - bugfix in transposing of RGB image data for coronal slice plots
+- voltage: decompress cbin
+  - add support for custom spikeglx.Reader
 
 ## Release Note 2.9